patching (not finished)
