+++ jwnotifyer/android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.teckinfor.jwnotifyer">
+    package="com.example.jwnotifyer">
    <application
         android:label="jwnotifyer"
         android:name="${applicationName}"
