+++ jwnotifyer/android/app/src/main/res/values/ic_launcher_background.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="ic_launcher_background">#6C8BC1</color>
+    <color name="ic_launcher_background">#3DDC84</color>
 </resources>
