+++ jwnotifyer/lib/store_data.dart
@@ -2,24 +2,27 @@ import 'dart:io';
 import 'dart:convert';
 import 'package:path_provider/path_provider.dart';
 
-class StoreData{
+class StoreData {
   StoreData();
 
-  void saveCurrentContext({required Map dataActiveLanguages, required Map dataSettings, required Map dataAvailableLanguages}){
+  void saveCurrentContext(
+      {required Map dataActiveLanguages,
+      required Map dataSettings,
+      required Map dataAvailableLanguages}) {
     saveActiveLanguages(dataActiveLanguages);
     saveSettings(dataSettings);
     saveAvailableLanguages(dataAvailableLanguages);
   }
 
-  void saveActiveLanguages(Map dataActiveLanguages){
+  void saveActiveLanguages(Map dataActiveLanguages) {
     writeData(_localFileActiveLanguages, dataActiveLanguages);
   }
 
-  void saveSettings(Map dataSettings){
-    writeData(_localFileSettings , dataSettings);
+  void saveSettings(Map dataSettings) {
+    writeData(_localFileSettings, dataSettings);
   }
 
-  void saveAvailableLanguages(Map dataAvailableLanguages){
+  void saveAvailableLanguages(Map dataAvailableLanguages) {
     writeData(_localFileAvailableLanguages, dataAvailableLanguages);
   }
 
@@ -73,10 +76,9 @@ class StoreData{
     try {
       final file = await localFile;
       final data = await file.readAsString();
-      return jsonDecode(data) ?? {"status" : "ERROR"};
+      return jsonDecode(data) ?? {"status": "ERROR"};
     } catch (e) {
       return {"status": "ERROR"};
     }
   }
-
 }
